feat(cmd): move subcommands to main; nit
